Try prepending path to a tag before trying as-is
If a repo (like https://github.com/aws/aws-sdk-go-v2) has too many tags, doing a lookup of _all_ tags results in a 502.

In a number of cases I found that the path is actually part of the tag, so we can try that to see if we can find the tag and reduce the number of times we might try getting all tags.